+++ Headers/Asservissement_HL.h
@@ -8,11 +8,11 @@
 void Mode_Asserv(int mode);
 int get_Mode_Asserv();
 extern signed long Asserv_Cons_distance;
-extern long double Asserv_Cons_angle;
-extern long double Asserv_zero;
-extern long double Asserv_dst_act;
+extern double Asserv_Cons_angle;
+extern double Asserv_zero;
+extern double Asserv_dst_act;
 void Gestion_Asserv_HL(long Tick_droit,long Tick_gauche,long* ordre_distance,long* ordre_angle);
 void Asserv_Reset_Integrateur(void);
 void Asserv_Reset_Derivateur(void);
-void set_Asserv_angle_abs(long double angle_abs);
+void set_Asserv_angle_abs(double angle_abs);
 #endif /* ASSERVISSEMENT_HL_H_ */
